flask_api_course/requirements.txt
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
passlib
marshmallow
python-dotenv
flask-migrate
gunicorn
redis
rq
psycopg2
requests
flask-cors